requirements.txt
fastapi[standard]
pydantic
google-generativeai
requests
beautifulsoup4
pypdf
pinecone

boto3
langchain
langchain-core
paramiko
PyJWT
redis
supabase
python-dotenv
langchain-google-genai
langchain-mistralai
langchain-community
langchain-google-vertexai
openai
tavily-python
sse-starlette
slack-sdk